Split k8s.io/component-base/registry and add into k8s.io/component-base/version and k8s.io/component-base/featuregate
