+++ pkg/controlplane/apiserver/options/options_test.go
@@ -37,13 +37,13 @@ import (
 	"k8s.io/component-base/featuregate"
 	"k8s.io/component-base/logs"
 	"k8s.io/component-base/metrics"
-	utilversion "k8s.io/component-base/registry"
+	utilversion "k8s.io/component-base/version"
 	kubeoptions "k8s.io/kubernetes/pkg/kubeapiserver/options"
 	netutils "k8s.io/utils/net"
 )
 
 func TestAddFlags(t *testing.T) {
-	componentGlobalsRegistry := utilversion.DefaultComponentGlobalsRegistry
+	componentGlobalsRegistry := featuregate.DefaultComponentGlobalsRegistry
 	t.Cleanup(func() {
 		componentGlobalsRegistry.Reset()
 	})
@@ -133,7 +133,7 @@ func TestAddFlags(t *testing.T) {
 			JSONPatchMaxCopyBytes:        int64(3 * 1024 * 1024),
 			MaxRequestBodyBytes:          int64(3 * 1024 * 1024),
 			ComponentGlobalsRegistry:     componentGlobalsRegistry,
-			ComponentName:                utilversion.DefaultKubeComponent,
+			ComponentName:                featuregate.DefaultKubeComponent,
 		},
 		Admission: &kubeoptions.AdmissionOptions{
 			GenericAdmission: &apiserveroptions.AdmissionOptions{
